Move `LeaderboardFlow` to `HUDOverlay` to share positioning logic
